backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
httpx==0.28.1
langchain==0.3.14
langchain-openai==0.3.0
langchain-community==0.3.14
langchain-text-splitters==0.3.4
faiss-cpu==1.9.0.post1
pypdf==5.1.0
openai==1.58.1
pydantic==2.10.4
